+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [..] unreleased
+### Changed
+- removed unnecessary columns in cluster mapping table [#45](https://github.com/graphsense/graphsense-tagpack-tool/issues/45)
+
 ## [1.1.0] 2022-10-11
 ### Added
 - Support for connection pooling